refactor: rename RunAllTriggers to TriggerAll for better semantic clarity
