+++ web/app/components/base/icons/src/vender/workflow/TriggerAll.json
@@ -69,5 +69,5 @@
 			}
 		]
 	},
-	"name": "RunAllTriggers"
+	"name": "TriggerAll"
 }
