+++ web/app/components/base/icons/src/vender/workflow/index.ts
@@ -19,9 +19,9 @@ export { default as LoopEnd } from './LoopEnd'
 export { default as Loop } from './Loop'
 export { default as ParameterExtractor } from './ParameterExtractor'
 export { default as QuestionClassifier } from './QuestionClassifier'
-export { default as RunAllTriggers } from './RunAllTriggers'
 export { default as Schedule } from './Schedule'
 export { default as TemplatingTransform } from './TemplatingTransform'
+export { default as TriggerAll } from './TriggerAll'
 export { default as VariableX } from './VariableX'
 export { default as WebhookLine } from './WebhookLine'
 export { default as WindowCursor } from './WindowCursor'
